FH modified plotPlate function and readCytoSet can now deal with empty FCS files. Floating point data example for readFCS
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/prada@11122 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ R/readFCS.R
@@ -70,7 +70,7 @@ readFCSdata <- function(con, offsets, x) {
 
   dattype <- switch(readFCSgetPar(x, "$DATATYPE"),
 	"I" = "integer",
-	"F" = "double",
+	"F" = "numeric",
     	stop(paste("Don't know how to deal with $DATATYPE", readFCSgetPar(x, "$DATATYPE")))) 
   
   if (readFCSgetPar(x, "$MODE") != "L")
